# Java/spring/basic/src/test/openapi/openapi3.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"tags":[{"name":"Plain Entity Example","description":"Plain Entity Example controller API"},{"name":"Simple","description":"Simple controller API (CRUD)"},{"name":"Plain Example","description":"Plain Example controller API"}],"paths":{"/simple/{id}":{"get":{"tags":["Simple"],"operationId":"getOne","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}}}}},"put":{"tags":["Simple"],"operationId":"updateOne","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}}}}},"delete":{"tags":["Simple"],"operationId":"deleteOne","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["Simple"],"operationId":"partialUpdateOne","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}}}}}},"/plain_entity/{id}":{"get":{"tags":["Plain Entity Example"],"summary":"Read a example item","description":"Read example item by index(id).","operationId":"readexample","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"string"}}}}}},"put":{"tags":["Plain Entity Example"],"summary":"Update example item","description":"Update item in the list where index indicate.","operationId":"updateexample","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"description":"Item to update","content":{"application/json":{"schema":{"type":"string"},"example":"coconut"}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["Plain Entity Example"],"summary":"Delete example item","description":"Delete item in the list where index indicate.","operationId":"deleteexample","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["Plain Entity Example"],"summary":"Partial update example item","description":"Update item in the list where index indicate (as partial).","operationId":"partialupdateexample","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"description":"Item to partialy update","content":{"application/json":{"schema":{"type":"string"},"example":"coconut"}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/plain/{id}":{"get":{"tags":["Plain Example"],"summary":"Read a example item","description":"Read example item by index(id).","operationId":"readexample_1","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"string"}}}}}},"put":{"tags":["Plain Example"],"summary":"Update example item","description":"Update item in the list where index indicate.","operationId":"updateexample_1","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"description":"Item to update","content":{"application/json":{"schema":{"type":"string"},"example":"coconut"}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["Plain Example"],"summary":"Delete example item","description":"Delete item in the list where index indicate.","operationId":"deleteexample_1","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"500":{"description":"Internal Server Error"},"200":{"description":"OK"},"404":{"description":"Not found"},"204":{"description":"No Content"}}},"patch":{"tags":["Plain Example"],"summary":"Partial update example item","description":"Update item in the list where index indicate (as partial).","operationId":"partialupdateexample_1","parameters":[{"name":"id","in":"path","description":"Index of example","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"description":"Item to partialy update","content":{"application/json":{"schema":{"type":"string"},"example":"coconut"}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/simple":{"get":{"tags":["Simple"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SimpleDto"}}}}}}},"post":{"tags":["Simple"],"operationId":"createOne","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleDto"}}}}}}},"/plain_entity":{"get":{"tags":["Plain Entity Example"],"summary":"Read example item list","description":"Read all of the example items","operationId":"allexample","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}},"post":{"tags":["Plain Entity Example"],"summary":"Create example item","description":"Create new item in the list.","operationId":"createexample","requestBody":{"description":"Item to create","content":{"application/json":{"schema":{"type":"string"},"example":"coconut"}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/plain":{"get":{"tags":["Plain Example"],"summary":"Read example item list","description":"Read all of the example items","operationId":"allexample_1","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}},"post":{"tags":["Plain Example"],"summary":"Create example item","description":"Create new item in the list.","operationId":"createexample_1","requestBody":{"description":"Item to create","content":{"application/json":{"schema":{"type":"string"},"example":"coconut"}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/simple/search":{"get":{"tags":["Simple"],"operationId":"getSearch","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageSimpleDto"}}}}}}},"/plain_entity/search":{"get":{"tags":["Plain Entity Example"],"summary":"Search for example item list","description":"Pagable search for retrieving paged item lists.","operationId":"searchexample","parameters":[{"name":"page","in":"query","description":"Current Page","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":1},{"name":"size","in":"query","description":"Page size","required":false,"schema":{"type":"integer","format":"int32","default":10},"example":10}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageString"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageString"}}}}}}},"/plain/search":{"get":{"tags":["Plain Example"],"summary":"Search for example item list","description":"Pagable search for retrieving paged item lists.","operationId":"searchexample_1","parameters":[{"name":"page","in":"query","description":"Current Page","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":1},{"name":"size","in":"query","description":"Page size","required":false,"schema":{"type":"integer","format":"int32","default":10},"example":10}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageString"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageString"}}}}}}}},"components":{"schemas":{"SimpleDto":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","description":"Simple record id (0 for new record)","format":"int32","example":1},"name":{"maxLength":50,"minLength":3,"type":"string","description":"Simple record name","example":"apple"}}},"PageSimpleDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/SimpleDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"PageString":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"type":"string"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}}}}